Move the compute function to it's own file
